refactor(edit-server.component.ts): Moved edit server to the server directory and changed it into a standalone component
