+++ src/BurstChat.Client/src/app/components/server/edit-server/edit-server.component.scss
@@ -1,0 +1,1 @@
+@import '../../../../scss/theme';
