--- src/BurstChat.Client/src/app/modules/servers/components/edit-server/edit-server.component.html
@@ -1,25 +1,0 @@
-<div class="burst-container-fluid bg-secondary" *ngIf="server">
-
-    <burst-card>
-
-        <burst-card-body>
-
-            <burst-edit-server-info [server]="server"></burst-edit-server-info>
-
-        </burst-card-body>
-
-        <burst-card-body>
-
-            <burst-edit-server-channels [server]="server"></burst-edit-server-channels>
-
-        </burst-card-body>
-
-        <burst-card-body>
-
-            <burst-edit-server-users [server]="server"></burst-edit-server-users>
-
-        </burst-card-body>
-
-    </burst-card>
-
-</div>
