--- src/BurstChat.Client/src/app/modules/servers/components/edit-server/edit-server.component.scss
@@ -1,1 +1,0 @@
-@import '../../../../../scss/theme';
