+++ src/BurstChat.Client/src/app/modules/servers/servers.module.ts
@@ -3,7 +3,6 @@ import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { FormsModule } from '@angular/forms';
 import { ServersRoutingModule } from 'src/app/modules/servers/servers.routing';
-import { EditServerComponent } from 'src/app/modules/servers/components/edit-server/edit-server.component';
 import { CardComponent } from 'src/app/components/shared/card/card.component';
 import { CardHeaderComponent } from 'src/app/components/shared/card-header/card-header.component';
 import { CardBodyComponent } from 'src/app/components/shared/card-body/card-body.component';
@@ -14,6 +13,7 @@ import { AvatarComponent } from 'src/app/components/shared/avatar/avatar.compone
 import { EditServerUsersComponent } from 'src/app/components/server/edit-server-users/edit-server-users.component';
 import { EditServerInfoComponent } from 'src/app/components/server/edit-server-info/edit-server-info.component';
 import { EditServerChannelsComponent } from 'src/app/components/server/edit-server-channels/edit-server-channels.component';
+import { EditServerComponent } from 'src/app/components/server/edit-server/edit-server.component';
 
 @NgModule({
     imports: [
@@ -30,11 +30,8 @@ import { EditServerChannelsComponent } from 'src/app/components/server/edit-serv
         AvatarComponent,
         EditServerUsersComponent,
         EditServerInfoComponent,
-        EditServerChannelsComponent
-    ],
-    declarations: [
-        EditServerComponent,
-    ],
-    providers: []
+        EditServerChannelsComponent,
+        EditServerComponent
+    ]
 })
 export class ServersModule { }
